# openapi.yaml
...
openapi: 3.1.0
info:
  title: Mapvest API
  version: 0.1.0-alpha.0
  description: Mapvest turns places and objects into investable tickers. All request and response shapes are generated from zod schemas in `packages/core/src/schemas` — never hand-edit `openapi.yaml`.
  contact:
    name: Mapvest
    url: https://github.com/jawauntb/mapvest
  license:
    name: UNLICENSED
servers:
  - url: http://localhost:3001
    description: Local dev
  - url: https://mapvest-api.up.railway.app
    description: Railway (staging/production)
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: "Session token issued by `POST /v1/auth/session`. Attach as `Authorization: Bearer <token>`."
  schemas:
...
    NearbyResponse:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
            properties:
              place:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  location:
                    type: object
                    properties:
                      lat:
                        type: number
                        minimum: -90
                        maximum: 90
                      lng:
                        type: number
                        minimum: -180
                        maximum: 180
                    required:
                      - lat
                      - lng
                  types:
                    type: array
                    items:
                      type: string
                    default: []
                required:
                  - id
                  - name
                  - location
              investable:
                type: object
                properties:
                  brand:
                    type: object
                    properties:
                      name:
                        type: string
                      parent:
                        type: string
                      isPublic:
                        type: boolean
                      ticker:
                        type: object
                        properties:
                          symbol:
                            type: string
                          exchange:
                            type: string
                          parent:
                            type: string
                        required:
                          - symbol
                      sector:
                        type: string
                      logo:
                        type: string
                        format: uri
                    required:
                      - name
                      - isPublic
                  comparables:
                    type: array
                    items:
                      type: object
                      properties:
                        ticker:
                          type: string
                        name:
                          type: string
                        score:
                          type: number
                          minimum: 0
                          maximum: 1
                        reasoning:
                          type: string
                        sources:
                          type: array
                          items:
                            type: object
                            properties:
                              provider:
                                type: string
                                enum:
                                  - exa
                                  - openrouter
                                  - gemini
                                  - massive
                                  - yahoo
                                  - polygon
                                  - sec
                                  - fred
                                  - manual
                              url:
                                type: string
                                format: uri
                              fetchedAt:
                                type: string
                              confidence:
                                type: string
                                enum:
                                  - high
                                  - medium
                                  - low
                            required:
                              - provider
                              - fetchedAt
                              - confidence
                      required:
                        - ticker
                        - name
                        - score
                        - reasoning
                        - sources
                    default: []
                  etfs:
                    type: array
                    items:
                      type: object
                      properties:
                        ticker:
                          type: string
                        name:
                          type: string
                        weight:
                          type: number
                          minimum: 0
                          maximum: 1
                        source:
                          type: object
                          properties:
                            provider:
                              type: string
                              enum:
                                - exa
                                - openrouter
                                - gemini
                                - massive
                                - yahoo
                                - polygon
                                - sec
                                - fred
                                - manual
                            url:
                              type: string
                              format: uri
                            fetchedAt:
                              type: string
                            confidence:
                              type: string
                              enum:
                                - high
                                - medium
                                - low
                          required:
                            - provider
                            - fetchedAt
                            - confidence
                      required:
                        - ticker
                        - name
                        - weight
                        - source
                    default: []
                  confidence:
                    type: string
                    enum:
                      - high
                      - medium
                      - low
                  sources:
                    type: array
                    items:
                      type: object
                      properties:
                        provider:
                          type: string
                          enum:
                            - exa
                            - openrouter
                            - gemini
                            - massive
                            - yahoo
                            - polygon
                            - sec
                            - fred
                            - manual
                        url:
                          type: string
                          format: uri
                        fetchedAt:
                          type: string
                        confidence:
                          type: string
                          enum:
                            - high
                            - medium
                            - low
                      required:
                        - provider
                        - fetchedAt
                        - confidence
                  quote:
                    type: object
                    properties:
                      symbol:
                        type: string
                      price:
                        type: number
                      change:
                        type: number
                      changePct:
                        type: number
                      currency:
                        type: string
                      ts:
                        type: string
                      disclaimer:
                        type: string
                      name:
                        type: string
                      provider:
                        type: string
                        enum:
                          - massive
                          - yahoo
                      freshness:
                        type: string
                        enum:
                          - real-time
                          - delayed
                          - end-of-day
                          - unknown
                    required:
                      - symbol
                      - price
                      - change
                      - changePct
                      - currency
                      - ts
                      - disclaimer
                required:
                  - brand
                  - confidence
                  - sources
            required:
              - place
      required:
        - items
...
    QuotaExceeded:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
          enum:
            - quota_exceeded
        remaining:
          type: integer
          minimum: 0
        limit:
          type: integer
          minimum: 0
        priceUsd:
          type: number
        interval:
          type: string
          enum:
            - month
      required:
        - error
        - code
        - remaining
        - limit
...
    TerritoryResponse:
      type: object
      properties:
        tile:
          type: string
        investablesTotal:
          type: number
        found:
          type: number
        pioneer:
          type: boolean
        sources:
          type: array
          items:
            type: object
            properties:
              provider:
                type: string
                enum:
                  - exa
                  - openrouter
                  - gemini
                  - massive
                  - yahoo
                  - polygon
                  - sec
                  - fred
                  - manual
              url:
                type: string
                format: uri
              fetchedAt:
                type: string
              confidence:
                type: string
                enum:
                  - high
                  - medium
                  - low
            required:
              - provider
              - fetchedAt
              - confidence
      required:
        - tile
        - investablesTotal
        - found
        - pioneer
        - sources
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details: {}
          required:
            - code
            - message
      required:
        - error
...
paths:
...
  /v1/territory:
    get:
      summary: Tile completion for a location
      description: Completion state for the geohash-6 tile containing `lat`/`lng`. `investablesTotal` is how many investable brands the same nearby cascade behind `/v1/nearby` resolves inside the tile, and `found` is how many of those the caller has already caught, so the two reconcile with `/v1/nearby` for that tile. `pioneer` is true when the caller has not yet recorded a find in this tile — the write-time pioneer XP bonus is still available here. The counts come from a live places + brand join, so the response carries `sources` (AGENTS.md §6) — an uncitable lookup returns fewer sources, never an invented one.
      tags:
        - identify
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type:
              - number
              - "null"
            minimum: -90
            maximum: 90
            example: 40.7128
          required: false
          name: lat
          in: query
        - schema:
            type:
              - number
              - "null"
            minimum: -180
            maximum: 180
            example: -74.006
          required: false
          name: lng
          in: query
      responses:
        "200":
          description: Tile completion resolved.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TerritoryResponse"
        "400":
          description: Bad request — validation failed against the zod schema.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid session token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "402":
          description: Free-tier generation quota spent. Clients must present the paywall, not a generic error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuotaExceeded"
        "429":
          description: Rate limit exceeded (per-user or per-ip).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /v1/nearby:
    get:
      summary: Nearby investable places
      description: "Google Places lookup around a lat/lng, filtered and annotated with tickers/ETFs. Every finance answer includes a `sources: Source[]` array (see AGENTS.md §6)."
      tags:
        - nearby
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: number
          required: true
          name: lat
          in: query
        - schema:
            type: number
          required: true
          name: lng
          in: query
        - schema:
            type: number
            default: 500
          required: false
          name: radius
          in: query
        - schema:
            type: number
            default: 25
          required: false
          name: limit
          in: query
      responses:
        "200":
          description: Nearby items resolved.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NearbyResponse"
        "400":
          description: Bad request — validation failed against the zod schema.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid session token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "402":
          description: Free-tier generation quota spent. Clients must present the paywall, not a generic error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuotaExceeded"
        "429":
          description: Rate limit exceeded (per-user or per-ip).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...